fixes #87
